algebra: separate libquat and libqrot from libvec3d
 - separate libquat as pure quaternion algebra library
 - separate libqrot as quaternion/vector rotations library
 - leave libvec3d as pure vector operations library
 - function naming unchanged

JIRA: PP-98
